Tag-Along-Backend/requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
Markdown==3.5.2
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
whitenoise==6.6.0
